requirements.txt
playwright
openai
python-dotenv
beautifulsoup4
Pillow
imagehash